requirements.txt
Django==1.8.4
MySQL-python==1.2.5
appnope==0.1.0
beautifulsoup4==4.4.0
decorator==4.0.2
django-extensions==1.5.6
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1
pexpect==3.3
pickleshare==0.5
python-dateutil==2.4.2
requests==2.7.0
simplegeneric==0.8.1
six==1.9.0
traitlets==4.0.0
wsgiref==0.1.2